\name{getFigureInfo}
\alias{getFigureInfo}

\title{
Get information about a figure in a Word document
}

\description{

This function retrieves meta information about a figure in a Word document.
This includes information such as 
     
}

\usage{
getFigureInfo(node, ar, 
               rid = getNodeSet(node, ".//ve:Choice//@r:embed", WordXMLNamespaces[c("ve", "r")]))
}

\arguments{


\item{node}{
}

\item{ar}{
}

\item{rid}{
}
}

\value{


}

\references{

}

\seealso{

}

  \author{Duncan Temple Lang \email{duncan@wald.ucdavis.edu}
}
 
\examples{
  
}

\keyword{

}
