\name{getStyleId}
\alias{getStyleId}
\alias{getStyles}
\alias{addStyle}

\title{
Get information about or add a style to a Word document
}

\description{

These functions allow us to access the collection of style definitions in a Word document
     and to add new style definitions.
     
}

\usage{
getStyleId(node)
getStyles(doc, useIds = FALSE)
addStyle(doc, style)
}

\arguments{


\item{doc}{
the Word document object as a WordArchive object.}

\item{useIds}{
}

\item{node}{
}

\item{style}{
the style object to be added. }
}

\value{


}

\references{

}

\seealso{

}

  \author{Duncan Temple Lang \email{duncan@wald.ucdavis.edu}
}
 
\examples{
  
}

\keyword{

}
