
\name{hyperlinks}
\alias{hyperlinks}
\alias{hyperlinks,WordArchive-method}
\alias{hyperlinks,WordArchive-method}

\title{Get the hyperlinks in a Word document}

\description{This function extracts the hyperlinks in a Word document, returning the
URLs and the text shown in the links.
     }

\usage{
hyperlinks(doc, comments = FALSE, ...)
}

\arguments{\item{doc}{}\item{comments}{}\item{\dots}{}}

\value{}

\references{}

\seealso{}

  \author{Duncan Temple Lang \email{duncan@r-project.org}
}
 
\examples{
  
}

\keyword{}
