
% Generated via xml2Rd.xsl from ./isParagraphNode.xml
\name{isParagraphNode}
\alias{isParagraphNode}

\title{
Determine if an XML node within a Word document represents a paragraph
}

\description{

 This function determines if the given XML node is the top-level node that 
      embodies a paragraph in the WordProcessingML structure.
     
}

\usage{
isParagraphNode(node)
}

\arguments{


\item{node}{
the XMLInternalNode}
}

\value{

A logical value indicating whether  the node represents a paragraph or not.
     
}

\references{

}

\seealso{

}

  \author{Duncan Temple Lang \email{duncan@wald.ucdavis.edu}
}
 
\examples{
  
}

\keyword{
programming
}
